// src/api/generated/models/fileNode.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FileNodeType } from './fileNodeType';

export interface FileNode {
  title?: string;
  path?: string;
  type?: FileNodeType;
  children?: FileNode[];
  content?: string;
}